requirements.txt
django
djangorestframework
markdown
psycopg2
django-filter
coreapi
gunicorn
requests
-e .